# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: JL Business Growth API
  version: 0.2.0
  description: HTTP contract for the JL团队生意成长管理系统.
servers:
  - url: /
...
paths:
...
  /api/finance/categories:
    get:
      operationId: listFinanceCategories
      tags: [finance-income]
      summary: List system and user finance categories
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Finance categories
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FinanceCategoryListResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
    post:
      operationId: createFinanceCategory
      tags: [finance-income]
      summary: Create a finance category
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FinanceCategoryRequest" }
      responses:
        "201":
          description: Finance category created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FinanceCategoryResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
  /api/finance/categories/{category_id}:
    delete:
      operationId: archiveFinanceCategory
      tags: [finance-income]
      summary: Archive a user finance category
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/CategoryId" }]
      responses:
        "204": { description: Finance category archived }
        default: { $ref: "#/components/responses/ErrorResponse" }
  /api/finance/transactions:
    get:
      operationId: listFinanceTransactions
      tags: [finance-income]
      summary: List financial transactions
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/DateFrom"
        - $ref: "#/components/parameters/DateTo"
      responses:
        "200":
          description: Financial transactions
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FinanceTransactionListResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
    post:
      operationId: createFinanceTransaction
      tags: [finance-income]
      summary: Create a financial transaction
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FinanceTransactionRequest" }
      responses:
        "201":
          description: Financial transaction created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FinanceTransactionResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
  /api/finance/transactions/{transaction_id}:
    put:
      operationId: updateFinanceTransaction
      tags: [finance-income]
      summary: Update a financial transaction
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/TransactionId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FinanceTransactionRequest" }
      responses:
        "200":
          description: Financial transaction updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FinanceTransactionResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deleteFinanceTransaction
      tags: [finance-income]
      summary: Delete a financial transaction
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/TransactionId" }]
      responses:
        "204": { description: Financial transaction deleted }
        default: { $ref: "#/components/responses/ErrorResponse" }
  /api/finance/budgets:
    get:
      operationId: listFinanceBudgets
      tags: [finance-income]
      summary: List monthly budgets
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Budgets
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FinanceBudgetListResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
    post:
      operationId: saveFinanceBudget
      tags: [finance-income]
      summary: Create or replace a monthly budget
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FinanceBudgetRequest" }
      responses:
        "200":
          description: Budget saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FinanceBudgetResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
  /api/finance/snapshots:
    get:
      operationId: listFinanceSnapshots
      tags: [finance-income]
      summary: List savings and emergency-fund snapshots
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Financial snapshots
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FinanceSnapshotListResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
    post:
      operationId: saveFinanceSnapshot
      tags: [finance-income]
      summary: Create or replace a financial snapshot
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FinanceSnapshotRequest" }
      responses:
        "200":
          description: Snapshot saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FinanceSnapshotResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
  /api/income-simulator/calculate:
    post:
      operationId: calculateIncome
      tags: [finance-income]
      summary: Calculate income using the Excel-compatible V1 rules without saving
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IncomeSimulationInput" }
      responses:
        "200":
          description: Income calculation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IncomeCalculationResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
  /api/income-simulations:
    get:
      operationId: listIncomeSimulations
      tags: [finance-income]
      summary: List saved income simulations
      security: [{ sessionCookie: [] }]
      responses:
        "200":
          description: Saved simulations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IncomeSimulationListResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
    post:
      operationId: createIncomeSimulation
      tags: [finance-income]
      summary: Save an income simulation snapshot
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IncomeSimulationRequest" }
      responses:
        "201":
          description: Simulation saved
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IncomeSimulationResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
  /api/income-simulations/compare:
    post:
      operationId: compareIncomeSimulations
      tags: [finance-income]
      summary: Compare saved income simulations
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ids]
              properties:
                ids: { type: array, minItems: 1, maxItems: 10, items: { type: string, format: uuid } }
      responses:
        "200":
          description: Simulation comparison
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IncomeSimulationListResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
  /api/income-simulations/{simulation_id}:
    get:
      operationId: getIncomeSimulation
      tags: [finance-income]
      summary: Get a saved income simulation
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/SimulationId" }]
      responses:
        "200":
          description: Saved simulation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IncomeSimulationResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
    put:
      operationId: updateIncomeSimulation
      tags: [finance-income]
      summary: Update a saved income simulation
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/SimulationId" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/IncomeSimulationRequest" }
      responses:
        "200":
          description: Simulation updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IncomeSimulationResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
    delete:
      operationId: deleteIncomeSimulation
      tags: [finance-income]
      summary: Delete a saved income simulation
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/SimulationId" }]
      responses:
        "204": { description: Simulation deleted }
        default: { $ref: "#/components/responses/ErrorResponse" }
  /api/income-simulations/{simulation_id}/duplicate:
    post:
      operationId: duplicateIncomeSimulation
      tags: [finance-income]
      summary: Duplicate a saved income simulation
      security: [{ sessionCookie: [] }]
      parameters: [{ $ref: "#/components/parameters/SimulationId" }]
      responses:
        "201":
          description: Simulation duplicated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/IncomeSimulationResponse" }
        default: { $ref: "#/components/responses/ErrorResponse" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-bos_session
  parameters:
...
    DateFrom:
      name: from
      in: query
      schema:
        type: string
        format: date
    DateTo:
      name: to
      in: query
      schema:
        type: string
        format: date
...
    CategoryId:
      name: category_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    TransactionId:
      name: transaction_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    SimulationId:
      name: simulation_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    ErrorResponse:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required: [error, request_id]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
        request_id:
          type: string
...
    FinanceCategory:
      type: object
      required: [id, type, name]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [INCOME, EXPENSE] }
        name: { type: string }
        archived_at: { type: string, format: date-time, nullable: true }
        system_default: { type: boolean }
    FinanceCategoryRequest:
      type: object
      required: [type, name]
      properties:
        type: { type: string, enum: [INCOME, EXPENSE] }
        name: { type: string, minLength: 1, maxLength: 100 }
    FinanceCategoryResponse:
      type: object
      required: [data, request_id]
      properties:
        data: { $ref: "#/components/schemas/FinanceCategory" }
        request_id: { type: string }
    FinanceCategoryListResponse:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/FinanceCategory" } }
        request_id: { type: string }
    FinanceTransaction:
      type: object
      required: [id, occurred_on, type, category_id, amount, source, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        occurred_on: { type: string, format: date }
        type: { type: string, enum: [INCOME, EXPENSE] }
        category_id: { type: string, format: uuid }
        amount: { type: string, pattern: '^\\d+(?:\\.\\d{1,2})?$' }
        description: { type: string, nullable: true }
        note: { type: string, nullable: true }
        source: { type: string, enum: [MANUAL, IMPORT] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    FinanceTransactionRequest:
      type: object
      required: [occurred_on, type, category_id, amount]
      properties:
        occurred_on: { type: string, format: date }
        type: { type: string, enum: [INCOME, EXPENSE] }
        category_id: { type: string, format: uuid }
        amount: { type: string, pattern: '^\\d+(?:\\.\\d{1,2})?$' }
        description: { type: string, nullable: true }
        note: { type: string, nullable: true }
        source: { type: string, enum: [MANUAL, IMPORT], default: MANUAL }
    FinanceTransactionResponse:
      type: object
      required: [data, request_id]
      properties:
        data: { $ref: "#/components/schemas/FinanceTransaction" }
        request_id: { type: string }
    FinanceTransactionListResponse:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/FinanceTransaction" } }
        request_id: { type: string }
    FinanceBudget:
      type: object
      required: [id, month, amount, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        month: { type: string, format: date }
        category_id: { type: string, format: uuid, nullable: true }
        amount: { type: string, pattern: '^\\d+(?:\\.\\d{1,2})?$' }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    FinanceBudgetRequest:
      type: object
      required: [month, amount]
      properties:
        month: { type: string, format: date }
        category_id: { type: string, format: uuid, nullable: true }
        amount: { type: string, pattern: '^\\d+(?:\\.\\d{1,2})?$' }
    FinanceBudgetResponse:
      type: object
      required: [data, request_id]
      properties:
        data: { $ref: "#/components/schemas/FinanceBudget" }
        request_id: { type: string }
    FinanceBudgetListResponse:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/FinanceBudget" } }
        request_id: { type: string }
    FinanceSnapshot:
      type: object
      required: [id, snapshot_date, kind, amount, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        snapshot_date: { type: string, format: date }
        kind: { type: string, enum: [SAVINGS, EMERGENCY_FUND] }
        amount: { type: string, pattern: '^\\d+(?:\\.\\d{1,2})?$' }
        note: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    FinanceSnapshotRequest:
      type: object
      required: [snapshot_date, kind, amount]
      properties:
        snapshot_date: { type: string, format: date }
        kind: { type: string, enum: [SAVINGS, EMERGENCY_FUND] }
        amount: { type: string, pattern: '^\\d+(?:\\.\\d{1,2})?$' }
        note: { type: string, nullable: true }
    FinanceSnapshotResponse:
      type: object
      required: [data, request_id]
      properties:
        data: { $ref: "#/components/schemas/FinanceSnapshot" }
        request_id: { type: string }
    FinanceSnapshotListResponse:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/FinanceSnapshot" } }
        request_id: { type: string }
    IncomeSimulationInput:
      type: object
      required: [personal_use_pv, customer_pv, markets, annual_growth_status, annual_growth_qualified_months, bfi_period_eligible, bbi_period_eligible, double_year_mode]
      properties:
        personal_use_pv: { type: number, minimum: 0 }
        customer_pv: { type: number, minimum: 0 }
        markets: { type: array, minItems: 12, maxItems: 12, items: { type: number, minimum: 0 } }
        annual_growth_status: { type: string, enum: [KEEP, GROWTH, NOT_QUALIFIED] }
        annual_growth_qualified_months: { type: integer, minimum: 0, maximum: 12 }
        bfi_period_eligible: { type: boolean }
        bbi_period_eligible: { type: boolean }
        double_year_mode: { type: string, enum: [NONE, FIRST, REPEAT] }
        double_year_rank: { type: string, nullable: true }
...
    IncomeCalculationResponse:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [rule_version, input, result]
          properties:
            rule_version: { type: string }
            input: { $ref: "#/components/schemas/IncomeSimulationInput" }
            result: { $ref: "#/components/schemas/IncomeSimulationResult" }
        request_id: { type: string }
    IncomeSimulation:
      type: object
      required: [id, name, rule_version, input, result, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        rule_version: { type: string }
        input: { $ref: "#/components/schemas/IncomeSimulationInput" }
        result: { $ref: "#/components/schemas/IncomeSimulationResult" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    IncomeSimulationRequest:
      type: object
      required: [name, input]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        input: { $ref: "#/components/schemas/IncomeSimulationInput" }
    IncomeSimulationResponse:
      type: object
      required: [data, request_id]
      properties:
        data: { $ref: "#/components/schemas/IncomeSimulation" }
        request_id: { type: string }
    IncomeSimulationListResponse:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/IncomeSimulation" } }
        request_id: { type: string }